Renamed selectors/index.js -> selectors/redux.js
